+++ web/changelog.md
@@ -25,6 +25,7 @@ In the meantime, Botania for Minecraft 1.20.1 may still receive updates for bug 
 * Fix: Hyacidus wasting mana on immune mobs that aren't undead
 * Fix: Various functional flowers never sent client updates when they used mana to perform their function
 * Fix: Rod of the Bifrost consumes mana again (AidaNightcore)
+* Fix: Cooldown of Thermalily and Munchdew can no longer be skipped by breaking the flower before it goes into cooldown
 * API: The `BlockProvider` interface now defines a method that lets others find out what the provider's "primary" block
   is. This is meant to be used by implementers of features that don't look for one specific block type, but need to know
   what's currently available.